csv stream read, 타입 유추 기능 추가

+++ src/controllers/api/ApiServiceController.ts
@@ -100,7 +100,7 @@ export class ApiServiceController {
         })
         const applicationId = service.application.id;
         await getManager().transaction("SERIALIZABLE", async transactionalEntityManager => {
-          await transactionalEntityManager.remove(service.meta);
+          if(service.meta) await transactionalEntityManager.remove(service.meta);
           await transactionalEntityManager.remove(service);
         });
         resolve({
